Try 5 times if sync failed.

+++ ftp_sync/ftp_sync.sh
@@ -26,7 +26,7 @@ if [[ ${options["only-existing"]} -eq 1 ]]; then
         tmp_file="/tmp/ftp_sync_$$_$(($(date +'%s * 1000 + %-N / 1000000')))"
         lftp -c "${cmd} > ${tmp_file}"
         return_code=$?;
-        if [[ ! $return_code -eq 0 ]]; then 
+        if [[ $return_code -ne 0 ]]; then 
             rm -f "${options["sync-list-file"]}"
             echo "Failed to retrieve file list in remote folder '${options["remote-dir"]}' from ${options["ftp-url"]}"; 
             exit $return_code;
@@ -49,7 +49,7 @@ if [[ ${options["only-existing"]} -eq 1 ]]; then
         rm -f "${tmp_file}"
         if [[ "${options["sync-list-filter"]}" != "" ]]; then
             source "${options["sync-list-filter"]}" "${options["sync-list-file"]}"
-            return_code=$?; if [[ ! $return_code -eq 0 ]]; then exit $return_code; fi
+            return_code=$?; if [[ $return_code -ne 0 ]]; then exit $return_code; fi
         fi
     else
         if [[ "${options["_log-level"]}" -ge "${log_levels["info"]}" ]]; then
@@ -104,9 +104,9 @@ if [[ ${options["only-existing"]} -eq 1 ]]; then
         if [[ ! -a "${options["sync-dir"]}/${f}" ]]; then
 	    #create parent path if required
             parent_dir=`get_parent_path "${options["sync-dir"]}/${f}"`
-            return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "$parent_dir"; exit $return_code; fi
+            return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "$parent_dir"; exit $return_code; fi
 	    create_dir "${parent_dir}"
-            return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "$parent_dir"; exit $return_code; fi
+            return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "$parent_dir"; exit $return_code; fi
 
             touch "${options["sync-dir"]}/${f}"
             if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
@@ -118,43 +118,99 @@ fi
 if [[ ${options["not-sync"]} -eq 1 ]]; then
     exit 0
 fi
+
 #start to synchronize the file between bom and local
 if [[ ${options["_log-level"]} -ge ${log_levels["info"]} ]]; then
     echo "${files_length} files will be synchronized from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}'"
 fi
-if [[ "$options["remote-dir"]" == "" ]]; then
-    lftp -c "open -u ${options["user"]},${options["password"]} ${options["ftp-url"]} && lcd ${options["sync-dir"]} && mirror -c --recursion=always --no-perms --no-umask ${sync_options};" 
-else
-    lftp -c "open -u ${options["user"]},${options["password"]} ${options["ftp-url"]} && cd ${options["remote-dir"]} && lcd ${options["sync-dir"]} && mirror -P 5 -c --recursion=always --no-perms --no-umask ${sync_options};" 
-fi
-return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to synchronize the file from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}'."; exit $return_code; fi
+#declare the try times if sync failed, and the waiting time before next try
+waiting_times=(0 60 120 300 300)
+return_code=1
+for wait_time in ${waiting_times[@]}; do
+    if [[ $wait_time -gt 0 ]]; then
+        if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+            echo "Wait ${wait_time} seconds before try again."
+        fi
+        sleep $wait_time
+    fi
+
+    if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+        echo "Start to synchronize file from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}'"
+    fi
+
+    if [[ "$options["remote-dir"]" == "" ]]; then
+        lftp -c "open -u ${options["user"]},${options["password"]} ${options["ftp-url"]} && lcd ${options["sync-dir"]} && mirror -c --recursion=always --no-perms --no-umask ${sync_options};" 
+    else
+        lftp -c "open -u ${options["user"]},${options["password"]} ${options["ftp-url"]} && cd ${options["remote-dir"]} && lcd ${options["sync-dir"]} && mirror -P 5 -c --recursion=always --no-perms --no-umask ${sync_options};" 
+    fi
+    return_code=$?
+
+    if [[ $return_code -eq 0 ]]; then
+        if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+            echo "End to synchronize file from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}'"
+        fi
+        break
+    elif [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+        echo "Failed to synchronize the file from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}'."
+    fi
+done
+if [[ $return_code -ne 0 ]]; then  echo "Failed to synchronize the file from '${options["ftp-url"]}/${options["remote-dir"]}' to '${options["sync-dir"]}',process exits."; exit $return_code; fi
 
 #switch the workspace from data-dir to sync-dir
 if [[ ${options["_log-level"]} -ge ${log_levels["info"]} ]]; then
     echo "Swich the workspce to '${options["sync-dir"]}'"
 fi
+
 if [[ -a "${options["local-dir"]}" ]]; then
     rm -f "${options["local-dir"]}"
-    return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to remove symbolic link '${options["local-dir"]}'."; exit $return_code; fi
+    return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "Failed to remove symbolic link '${options["local-dir"]}'."; exit $return_code; fi
 fi
 ln -s "${options["sync-dir"]}" "${options["local-dir"]}"
-return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to create symbolic link '${options["local-dir"]}' to '${options["sync-dir"]}'."; exit $return_code; fi
+
+return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "Failed to create symbolic link '${options["local-dir"]}' to '${options["sync-dir"]}'."; exit $return_code; fi
 
 #start to synchronize the file between local and work dir
 if [[ ${options["_log-level"]} -ge ${log_levels["info"]} ]]; then
     echo "${files_length} files will be synchronized from '${options["sync-dir"]}' to '${options["data-dir"]}'"
 fi
-lftp -c "local mirror --no-perms --no-umask --recursion=always ${options["sync-dir"]} ${options["data-dir"]}"
-return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to synchronize the file from '${options["sync-dir"]}' to '${options["data-dir"]}'."; exit $return_code; fi
+waiting_times=(0 60 60 60 60)
+return_code=1
+for wait_time in ${waiting_times[@]}; do
+    if [[ $wait_time -gt 0 ]]; then
+        if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+            echo "Wait ${wait_time} seconds before try again."
+        fi
+        sleep $wait_time
+    fi
+
+    if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+        echo "Start to synchronize file from '${options["sync-dir"]}' to '${options["data-dir"]}'"
+    fi
+
+    lftp -c "local mirror --no-perms --no-umask --recursion=always ${options["sync-dir"]} ${options["data-dir"]}"
+    return_code=$?
+
+    if [[ $return_code -eq 0 ]]; then
+        if [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+            echo "End to synchronize file from '${options["sync-dir"]}' to '${options["data-dir"]}'"
+        fi
+        break
+    elif [[ ${options["_log-level"]} -ge ${log_levels["debug"]} ]]; then
+        echo "Failed to synchronize the file from '${options["sync-dir"]}' to '${options["data-dir"]}'."
+    fi
+done
+if [[ $return_code -ne 0 ]]; then echo "Failed to synchronize the file from '${options["sync-dir"]}' to '${options["data-dir"]}', process exits."; exit $return_code; fi
 
 #switch the workspace from sync-dir to data-dir
 if [[ ${options["_log-level"]} -ge ${log_levels["info"]} ]]; then
     echo "Swich the workspce to '${options["data-dir"]}'"
 fi
-rm -f "${options["local-dir"]}"
-return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to remove symbolic link '${options["local-dir"]}'."; exit $return_code; fi
+if [[ -a "${options["local-dir"]}" ]]; then
+    rm -f "${options["local-dir"]}"
+    return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "Failed to remove symbolic link '${options["local-dir"]}'."; exit $return_code; fi
+fi
 ln -s "${options["data-dir"]}" "${options["local-dir"]}"
-return_code=$?; if [[ ! $return_code -eq 0 ]]; then  echo "Failed to create symbolic link '${options["local-dir"]}' to '${options["data-dir"]}'."; exit $return_code; fi
+return_code=$?; if [[ $return_code -ne 0 ]]; then  echo "Failed to create symbolic link '${options["local-dir"]}' to '${options["data-dir"]}'."; exit $return_code; fi
 
 if [[ ${options["_log-level"]} -ge ${log_levels["info"]} ]]; then
     echo "End to synchronization at `date`"
